'update_mail'

+++ MotelBE/src/main/java/com/lvh/controllers/ApiCommentController.java
@@ -10,18 +10,16 @@ import com.lvh.services.CommentSecurityService;
 import com.lvh.services.CommentService;
 import java.security.Principal;
 import java.util.Map;
-import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -31,10 +29,9 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 @RequestMapping("/api/comments")
+@CrossOrigin(origins = {"*"})
 public class ApiCommentController {
 
-
-    
     @Autowired
     private CommentService cmtSer;
     
@@ -47,25 +44,22 @@ public class ApiCommentController {
     })
     @ResponseStatus(HttpStatus.CREATED)
     @CrossOrigin
-    public void create(@RequestParam Map<String, String> params) {
+    public void create(@RequestBody Map<String, String> params) {
         this.cmtSer.saveOrUpdateComment(this.cmtSer.jsonToComment(params));
     }
-
     
     @DeleteMapping("/{commentId}")
     public ResponseEntity<?> deleteMotel(@PathVariable(value="commentId") Long id, Principal p) {
         Comment comment = this.cmtSer.getCommentById(id);
+        if (comment.getUserId() ==  null)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
         String currentUsername = p.getName();
-        System.out.println("zzzzzzzzzzzzzzzzzz");
-        System.out.println(comment.getUserId().getUsername());
-        System.out.println("bbbbbbbbbb");
-        System.out.println(currentUsername);
         if (comment.getUserId().getUsername().equals(currentUsername)){
             this.cmtSer.deleteComment(id);
             return ResponseEntity.noContent().build();
-
         }
-        return ResponseEntity.badRequest().body(id);
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 
     }
 }
